rfctr(html): harvest usable HTMLDocument tests
Some of the HTMLDocument tests are useful and relevant as integration
tests for `partition_html()`. Others are implementation specific and are
replaced by unit tests on the new parser.

Salvage those that will be useful after swapping in the new parser and
add them to `test_unstructured/partition/html/test_partition.py`.

These will be adjusted after installing the new parser to call
`partition_html()` instead of `HTMLDocument` to get their elements.
